Improve instrument name validation and log messages
